Finished ProviderControllerOp!!!

+++ gitRepositoryTeam6/chocAnSystem/src/chocAnSystem/ProviderRecord.java
@@ -4,7 +4,7 @@ public class ProviderRecord extends RosterRecord {
     private double fee;
 
     // Constructor
-    public ProviderRecord(String name, int number, String address, String city, String state, int zip, double fee) {
+    public ProviderRecord(String name, long number, String address, String city, String state, int zip, double fee) {
         super(name, number, address, city, state, zip);
         this.fee = fee;
     }
